@@ -1,15 +1,18 @@
 apply plugin: 'com.android.application'
-
+repositories {
+    // Documentation doesn't specify this, but it must be added here
+//    google()
+}
 android {
-    compileSdkVersion 26
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 29
+    buildToolsVersion '26.0.3'
     defaultConfig {
         applicationId "james.metronome"
         minSdkVersion 19
-        targetSdkVersion 26
+        targetSdkVersion 29
         versionCode 15
         versionName "1.7"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         vectorDrawables.useSupportLibrary = true
     }
     buildTypes {
@@ -28,16 +31,18 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    androidTestCompile('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile 'com.android.support:appcompat-v7:26.1.0'
-    compile 'com.android.support:design:26.1.0'
-    compile 'com.github.afollestad:aesthetic:-SNAPSHOT'
-    compile 'com.afollestad.material-dialogs:core:0.9.4.7'
-    compile 'com.github.bumptech.glide:glide:4.1.1'
-    testCompile 'junit:junit:4.12'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'com.google.android.material:material:1.3.0'
+    implementation 'com.github.afollestad:aesthetic:-SNAPSHOT'
+    implementation 'com.afollestad.material-dialogs:core:0.9.6.0'
+    implementation 'com.github.bumptech.glide:glide:4.1.1'
+    implementation 'com.squareup.retrofit2:adapter-rxjava2:2.5.0'
+//    implementation 'com.google.android.support:wearable:2.8.1'
+    testImplementation 'junit:junit:4.12'
 }
 
 configurations.all {
