@@ -1,8 +1,8 @@
 package james.metronome.views;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.v4.view.GravityCompat;
+import androidx.annotation.Nullable;
+import androidx.core.view.GravityCompat;
 import android.util.AttributeSet;
 import android.view.Gravity;
 import android.view.View;
