@@ -1,7 +1,7 @@
 package james.metronome.views;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatImageView;
+import androidx.appcompat.widget.AppCompatImageView;
 import android.util.AttributeSet;
 
 public class RatioImageView extends AppCompatImageView {
