@@ -1,13 +1,12 @@
 package james.metronome.data;
 
 import android.content.Context;
-import android.support.annotation.ColorRes;
-import android.support.annotation.StringRes;
-import android.support.v4.content.ContextCompat;
 
 import com.afollestad.aesthetic.Aesthetic;
 
-import james.metronome.R;
+import androidx.annotation.ColorRes;
+import androidx.annotation.StringRes;
+import androidx.core.content.ContextCompat;
 import james.metronome.utils.ColorUtils;
 
 public class ThemeData {
@@ -33,13 +32,13 @@ public class ThemeData {
         boolean isBackgroundDark = ColorUtils.isColorDark(backgroundColor);
 
         Aesthetic.get()
-                .colorPrimary(ContextCompat.getColor(context, colorPrimaryRes))
-                .colorAccent(ContextCompat.getColor(context, colorAccentRes))
-                .colorWindowBackground(backgroundColor)
-                .textColorPrimary(ContextCompat.getColor(context, isBackgroundDark ? R.color.textColorPrimary : R.color.textColorPrimaryInverse))
-                .textColorPrimaryInverse(ContextCompat.getColor(context, isBackgroundDark ? R.color.textColorPrimaryInverse : R.color.textColorPrimary))
-                .textColorSecondary(ContextCompat.getColor(context, isBackgroundDark ? R.color.textColorSecondary : R.color.textColorSecondaryInverse))
-                .textColorSecondaryInverse(ContextCompat.getColor(context, isBackgroundDark ? R.color.textColorSecondaryInverse : R.color.textColorSecondary))
+//                .colorPrimary(ContextCompat.getColor(context, colorPrimaryRes))
+//                .colorAccent(ContextCompat.getColor(context, colorAccentRes))
+//                .colorWindowBackground(backgroundColor)
+//                .textColorPrimary(ContextCompat.getColor(context, isBackgroundDark ? R.color.textColorPrimary : R.color.textColorPrimaryInverse))
+//                .textColorPrimaryInverse(ContextCompat.getColor(context, isBackgroundDark ? R.color.textColorPrimaryInverse : R.color.textColorPrimary))
+//                .textColorSecondary(ContextCompat.getColor(context, isBackgroundDark ? R.color.textColorSecondary : R.color.textColorSecondaryInverse))
+//                .textColorSecondaryInverse(ContextCompat.getColor(context, isBackgroundDark ? R.color.textColorSecondaryInverse : R.color.textColorSecondary))
                 .colorStatusBarAuto()
                 .colorNavigationBarAuto()
                 .apply();
